requirements.txt
fastapi
uvicorn[standard]
langchain
psycopg2-binary
redis
python-dotenv
python-docx
apscheduler
sqlalchemy
pydantic
tqdm
typer[all]
sqlalchemy_utils